Decomposed simulator tests
